+++ src/BaseballGameView.js
@@ -1,78 +1,33 @@
-import { $ } from './utils/index.js';
-import { generateRandomNumber } from './computer/computer.js';
-import User from './user/user.js';
-import { BaseballGame } from './index.js';
-
-export default class BaseballGameView {
-  constructor({ $restartButton, $submit, $result, $correctResult }) {
-    this.user = new User();
-    this.baseballgame = new BaseballGame();
-    this.$restartButton = $restartButton;
-    this.$submit = $submit;
-    this.$result = $result;
-    this.$correctResult = $correctResult;
-    this.computerInputNumbers = generateRandomNumber();
-  }
-
-  init() {
-    this.triggerRestartEvent();
-    this.triggerSubmitEvent();
-    console.log(this.computerInputNumbers);
-  }
-
-  triggerSubmitEvent() {
-    this.$submit.addEventListener('click', (e) => this.onAnswerSubmit(e));
-  }
-
-  onAnswerSubmit(e) {
-    e.preventDefault();
-    if (!this.user.isInputValid(this.user.getUserInputValue())) {
-      return;
+export class BaseballGame {
+  play(computerInputNumbers, userInputNumbers) {
+    const ball = this.calculateBall(computerInputNumbers, userInputNumbers);
+    const strike = this.calculateStrike(computerInputNumbers, userInputNumbers);
+    if (ball || strike) {
+      return `${ball} ${strike}`;
     }
-    if (this.computerInputNumbers === this.user.getUserInputValue()) {
-      this.makeVisible('$correctResult');
-      return;
+    return '낫싱';
+  }
+
+  calculateBall(computerInputNumbers, userInputNumbers) {
+    let ball = 0;
+    for (let i = 0; i < computerInputNumbers.length; i++) {
+      if (
+        computerInputNumbers.includes(userInputNumbers[i]) &&
+        computerInputNumbers[i] !== userInputNumbers[i]
+      ) {
+        ball++;
+      }
     }
-    const count = this.baseballgame.play(this.computerInputNumbers, this.user.getUserInputValue());
-    this.$result.innerHTML = count;
+    return ball > 0 ? `${ball}볼` : '';
   }
 
-  triggerRestartEvent() {
-    this.$restartButton.addEventListener('click', (e) => this.onRestart(e));
-  }
-
-  onRestart(e) {
-    e.preventDefault();
-    this.user.setUserInputValue('');
-    this.makeVisible('$result');
-    this.computerInputNumbers = generateRandomNumber();
-  }
-
-  makeVisible($element) {
-    if ($element === '$result') {
-      this.showResult();
-    } else if ($element === '$correctResult') {
-      this.showCorrectResult();
+  calculateStrike(computerInputNumbers, userInputNumbers) {
+    let strike = 0;
+    for (let i = 0; i < computerInputNumbers.length; i++) {
+      if (computerInputNumbers[i] === userInputNumbers[i]) {
+        strike++;
+      }
     }
-  }
-
-  showResult() {
-    this.$result.innerText = '';
-    this.$result.style.display = 'block';
-    this.$correctResult.style.display = 'none';
-  }
-
-  showCorrectResult() {
-    this.$result.style.display = 'none';
-    this.$correctResult.style.display = 'block';
+    return strike > 0 ? `${strike}스트라이크` : '';
   }
 }
-
-const game = new BaseballGameView({
-  $restartButton: $('#game-restart-button'),
-  $correctResult: $('#submit'),
-  $result: $('#result'),
-  $submit: $('#correct-result'),
-});
-
-game.init();
